pre-delinquency-engine/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning
xgboost==2.0.2
scikit-learn==1.3.2
joblib==1.3.2

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Kafka
kafka-python==2.0.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# Data Generation
Faker==20.1.0

# Optional: MLflow (for full training)
# mlflow==2.9.2

# Optional: SHAP (for explainability)
# shap==0.43.0

# Optional: Dashboard (Phase 6)
# streamlit==1.29.0
# plotly==5.18.0

# Optional: Testing
# pytest==7.4.3
# pytest-cov==4.1.0
